EditorContext moved into top components folder + code cleanup
