+++ src/components/common/PageContent/PageContent.scss
@@ -1,4 +1,4 @@
-@import '../../../stylesheets/common.scss';
+@import '@stylesheets/common.scss';
 .page-content {
     padding: $spacing-xs $spacing $spacing $spacing;
     overflow: auto;
