+++ src/components/common/PageHeader/PageHeader.scss
@@ -1,4 +1,4 @@
-@import '../../../stylesheets/common.scss';
+@import '@stylesheets/common.scss';
 .header {
     display: flex;
     flex-direction: row;
